[🚚move] 회원 관리 기존 코드 백업
